+++ src/InMemoryRepository.php
@@ -11,8 +11,8 @@
 
 namespace Puli\Repository;
 
-use Assert\Assertion;
 use InvalidArgumentException;
+use Puli\Repository\Assert\Assertion;
 use Puli\Repository\Resource\Collection\ArrayResourceCollection;
 use Puli\Repository\Resource\Collection\ResourceCollection;
 use Puli\Repository\Resource\DirectoryResource;
@@ -94,9 +94,7 @@ class InMemoryRepository implements ManageableRepository
      */
     public function get($path)
     {
-        Assertion::string($path, 'The path must be a string. Got: %2$s');
-        Assertion::notEmpty($path, 'The path must not be empty.');
-        Assertion::startsWith($path, '/', 'The path %s is not absolute.');
+        Assertion::path($path);
 
         $path = Path::canonicalize($path);
 
@@ -112,9 +110,7 @@ class InMemoryRepository implements ManageableRepository
      */
     public function find($selector)
     {
-        Assertion::string($selector, 'The selector must be a string. Got: %2$s');
-        Assertion::notEmpty($selector, 'The selector must not be empty.');
-        Assertion::startsWith($selector, '/', 'The selector %s is not absolute.');
+        Assertion::selector($selector);
 
         $selector = Path::canonicalize($selector);
         $staticPrefix = Selector::getStaticPrefix($selector);
@@ -147,9 +143,7 @@ class InMemoryRepository implements ManageableRepository
      */
     public function contains($selector)
     {
-        Assertion::string($selector, 'The selector must be a string. Got: %2$s');
-        Assertion::notEmpty($selector, 'The selector must not be empty.');
-        Assertion::startsWith($selector, '/', 'The selector %s is not absolute.');
+        Assertion::selector($selector);
 
         $selector = Path::canonicalize($selector);
         $staticPrefix = Selector::getStaticPrefix($selector);
@@ -193,9 +187,7 @@ class InMemoryRepository implements ManageableRepository
      */
     public function add($path, $resource)
     {
-        Assertion::string($path, 'The path must be a string. Got: %2$s');
-        Assertion::notEmpty($path, 'The path must not be empty.');
-        Assertion::startsWith($path, '/', 'The path %s is not absolute.');
+        Assertion::path($path);
 
         $path = Path::canonicalize($path);
 
@@ -244,9 +236,7 @@ class InMemoryRepository implements ManageableRepository
      */
     public function remove($selector)
     {
-        Assertion::string($selector, 'The selector must be a string. Got: %2$s');
-        Assertion::notEmpty($selector, 'The selector must not be empty.');
-        Assertion::startsWith($selector, '/', 'The selector %s is not absolute.');
+        Assertion::selector($selector);
 
         $selector = Path::canonicalize($selector);
 
@@ -291,9 +281,7 @@ class InMemoryRepository implements ManageableRepository
      */
     public function listDirectory($path)
     {
-        Assertion::string($path, 'The path must be a string. Got: %2$s');
-        Assertion::notEmpty($path, 'The path must not be empty.');
-        Assertion::startsWith($path, '/', 'The path %s is not absolute.');
+        Assertion::path($path);
 
         $path = Path::canonicalize($path);